+++ lib/karafka/web/pro/ui/controllers/explorer/messages/exporting_controller.rb
@@ -1,0 +1,92 @@
+# frozen_string_literal: true
+
+# Karafka Pro - Source Available Commercial Software
+# Copyright (c) 2017-present Maciej Mensfeld. All rights reserved.
+#
+# This software is NOT open source. It is source-available commercial software
+# requiring a paid license for use. It is NOT covered by LGPL.
+#
+# The author retains all right, title, and interest in this software,
+# including all copyrights, patents, and other intellectual property rights.
+# No patent rights are granted under this license.
+#
+# PROHIBITED:
+# - Use without a valid commercial license
+# - Redistribution, modification, or derivative works without authorization
+# - Reverse engineering, decompilation, or disassembly of this software
+# - Use as training data for AI/ML models or inclusion in datasets
+# - Scraping, crawling, or automated collection for any purpose
+#
+# PERMITTED:
+# - Reading, referencing, and linking for personal or commercial use
+# - Runtime retrieval by AI assistants, coding agents, and RAG systems
+#   for the purpose of providing contextual help to Karafka users
+#
+# Receipt, viewing, or possession of this software does not convey or
+# imply any license or right beyond those expressly stated above.
+#
+# License: https://karafka.io/docs/Pro-License-Comm/
+# Contact: contact@karafka.io
+
+module Karafka
+  module Web
+    module Pro
+      module Ui
+        module Controllers
+          module Explorer
+            module Messages
+              # Streams an existing message payload to the browser as a file, either raw or as the
+              # deserialized-and-reserialized JSON export
+              class ExportingController < BaseController
+                # Dispatches the message raw payload to the browser as a file
+                #
+                # @param topic_id [String]
+                # @param partition_id [Integer]
+                # @param offset [Integer] offset of the message we want to download
+                def download(topic_id, partition_id, offset)
+                  message = Models::Message.find(topic_id, partition_id, offset)
+
+                  deny! unless visibility_filter.download?(message)
+
+                  file(
+                    message.raw_payload,
+                    "#{topic_id}_#{partition_id}_#{offset}_payload.msg"
+                  )
+                end
+
+                # Dispatches the message payload first deserialized and then serialized to JSON
+                # It differs from the raw payload in cases where raw payload is compressed or binary
+                # or contains data that the Web UI user should not see that was altered on the Web
+                # UI with the visibility filter.
+                #
+                # @param topic_id [String]
+                # @param partition_id [Integer]
+                # @param offset [Integer] offset of the message we want to export
+                def export(topic_id, partition_id, offset)
+                  Lib::PatternsDetector.new.call
+
+                  message = Models::Message.find(topic_id, partition_id, offset)
+
+                  # Check if exports are allowed
+                  deny! unless visibility_filter.export?(message)
+
+                  # Payload may deserialize correctly but still not be serializable back to JSON,
+                  # for example when it contains non-UTF-8 byte sequences. In such cases there is
+                  # no JSON representation that could be exported
+                  payload_json = Lib::SafeRunner.new { message.payload.to_json }.tap(&:call)
+
+                  not_found!(topic_id) unless payload_json.success?
+
+                  file(
+                    payload_json.result,
+                    "#{topic_id}_#{partition_id}_#{offset}_payload.json"
+                  )
+                end
+              end
+            end
+          end
+        end
+      end
+    end
+  end
+end
