+++ lib/karafka/web/pro/ui/views/explorer/messages/publishing/publish.erb
