+++ lib/karafka/web/pro/ui/views/explorer/messages/republishing/_breadcrumbs.erb
@@ -1,0 +1,49 @@
+<%# Karafka Pro - Source Available Commercial Software %>
+<%# Copyright (c) 2017-present Maciej Mensfeld. All rights reserved. %>
+<%# This software is NOT open source. It is source-available commercial software %>
+<%# requiring a paid license for use. It is NOT covered by LGPL. %>
+<%# The author retains all right, title, and interest in this software, %>
+<%# including all copyrights, patents, and other intellectual property rights. %>
+<%# No patent rights are granted under this license. %>
+<%# PROHIBITED: Use without a valid commercial license %>
+<%# PROHIBITED: Redistribution, modification, or derivative works without authorization %>
+<%# PROHIBITED: Reverse engineering, decompilation, or disassembly of this software %>
+<%# PROHIBITED: Use as training data for AI/ML models or inclusion in datasets %>
+<%# PROHIBITED: Scraping, crawling, or automated collection for any purpose %>
+<%# PERMITTED: Reading, referencing, and linking for personal or commercial use %>
+<%# PERMITTED: Runtime retrieval by AI assistants, coding agents, and RAG systems %>
+<%# PERMITTED: for the purpose of providing contextual help to Karafka users %>
+<%# Receipt, viewing, or possession of this software does not convey or %>
+<%# imply any license or right beyond those expressly stated above. %>
+<%# License: https://karafka.io/docs/Pro-License-Comm/ %>
+<%# Contact: contact@karafka.io %>
+
+<li>
+  <a href="<%= explorer_path %>">
+    Explorer
+  </a>
+</li>
+
+<li>
+  <a href="<%= explorer_topics_path(@topic_id) %>">
+    <%= @topic_id %>
+  </a>
+</li>
+
+<li>
+  <a href="<%= explorer_topics_path(@topic_id, @partition_id) %>">
+    Partition <%= @partition_id %>
+  </a>
+</li>
+
+<li>
+  <a href="<%= explorer_topics_path(@topic_id, @partition_id, @offset) %>">
+    <%= @offset %>
+  </a>
+</li>
+
+<li>
+  <a href="<%= explorer_messages_path(@topic_id, @partition_id, @offset, 'forward') %>">
+    Forward
+  </a>
+</li>
