+++ lib/karafka/web/pro/ui/views/explorer/messages/republishing/forward.erb
