+++ test/lib/karafka/web/pro/ui/controllers/explorer/messages/publishing_controller_test.rb
@@ -32,132 +32,6 @@ describe_current do
   let(:app) { Karafka::Web::Pro::Ui::App }
 
   let(:topic) { create_topic }
-  let(:target_topic) { create_topic }
-
-  describe "#forward" do
-    context "when we want to republish message from a non-existing topic" do
-      before { get "explorer/messages/non-existing/0/1/forward" }
-
-      it do
-        refute(response.ok?)
-        assert_equal(404, response.status)
-      end
-    end
-
-    context "when message exists" do
-      let(:payload) { rand.to_s }
-
-      before do
-        produce(topic, payload)
-        get "explorer/messages/#{topic}/0/0/forward"
-      end
-
-      it do
-        assert_ok
-        assert_body(topic)
-        assert_body("message-republish-form")
-        refute_body(pagination)
-      end
-    end
-  end
-
-  describe "#republish" do
-    context "when we want to republish message from a non-existing topic" do
-      before { post "explorer/messages/non-existing/0/1/republish" }
-
-      it do
-        refute(response.ok?)
-        assert_equal(404, response.status)
-      end
-    end
-
-    context "when message exists" do
-      let(:republished) { wait_for_message(target_topic, 0, 0) }
-      let(:payload) { rand.to_s }
-      let(:target_partition) { 0 }
-      let(:include_source_headers) { "on" }
-      let(:params) do
-        {
-          target_topic: target_topic,
-          target_partition: target_partition,
-          include_source_headers: include_source_headers
-        }
-      end
-
-      context "when we do not specify the target partition" do
-        let(:target_partition) { "" }
-
-        before do
-          produce(topic, payload)
-          post "explorer/messages/#{topic}/0/0/republish", params
-        end
-
-        it do
-          assert_equal(302, response.status)
-          # Taken from referer and referer is nil in specs
-          assert_equal("/", response.location)
-          assert_equal(payload, republished.raw_payload)
-          assert_includes(republished.headers.keys, "source_topic")
-          assert_includes(republished.headers.keys, "source_partition")
-          assert_includes(republished.headers.keys, "source_offset")
-        end
-      end
-
-      context "when we do not want source headers" do
-        let(:include_source_headers) { false }
-
-        before do
-          produce(topic, payload)
-          post "explorer/messages/#{topic}/0/0/republish", params
-        end
-
-        it do
-          assert_equal(302, response.status)
-          assert_equal("/", response.location)
-          assert_equal(payload, republished.raw_payload)
-          refute_includes(republished.headers.keys, "source_topic")
-          refute_includes(republished.headers.keys, "source_partition")
-          refute_includes(republished.headers.keys, "source_offset")
-        end
-      end
-
-      context "when we specify target partition" do
-        let(:target_partition) { 1 }
-        let(:republished) { wait_for_message(target_topic, 1, 0) }
-        let(:target_topic) { create_topic(partitions: 2) }
-
-        before do
-          produce(topic, payload)
-          post "explorer/messages/#{topic}/0/0/republish", params
-        end
-
-        it do
-          assert_equal(302, response.status)
-          assert_equal("/", response.location)
-          assert_equal(payload, republished.raw_payload)
-          assert_includes(republished.headers.keys, "source_topic")
-          assert_includes(republished.headers.keys, "source_partition")
-          assert_includes(republished.headers.keys, "source_offset")
-        end
-      end
-    end
-
-    context "when message exists but republishing is off" do
-      let(:payload) { rand.to_s }
-
-      before do
-        Karafka::Web.config.ui.policies.messages.stubs(:republish?).returns(false)
-
-        produce(topic, payload)
-        post "explorer/messages/#{topic}/0/0/republish"
-      end
-
-      it do
-        refute(response.ok?)
-        assert_equal(403, response.status)
-      end
-    end
-  end
 
   describe "#publish" do
     context "when the topic does not exist" do
@@ -419,143 +293,6 @@ describe_current do
 
       it "renders the error exactly once" do
         assert_equal(1, response.body.scan("could not be serialized").count)
-      end
-    end
-  end
-
-  describe "#download" do
-    context "when we want to download message from a non-existing topic" do
-      before { get "explorer/messages/non-existing/0/1/download" }
-
-      it do
-        refute(response.ok?)
-        assert_equal(404, response.status)
-      end
-    end
-
-    context "when message exists" do
-      let(:payload) { rand.to_s }
-      let(:expected_file_name) { "#{topic}_0_0_payload.msg" }
-      let(:expected_disposition) { "attachment; filename=\"#{expected_file_name}\"" }
-
-      before do
-        produce(topic, payload)
-        get "explorer/messages/#{topic}/0/0/download"
-      end
-
-      it do
-        assert_ok
-        assert_equal(expected_disposition, response.headers["content-disposition"])
-        assert_equal("application/octet-stream", response.headers["content-type"])
-        assert_equal(payload, response.body)
-      end
-    end
-
-    context "when message exists but downloads are off" do
-      let(:payload) { rand.to_s }
-
-      before do
-        Karafka::Web.config.ui.policies.messages.stubs(:download?).returns(false)
-
-        produce(topic, payload)
-        get "explorer/messages/#{topic}/0/0/download"
-      end
-
-      it do
-        refute(response.ok?)
-        assert_equal(403, response.status)
-      end
-    end
-  end
-
-  describe "#export" do
-    context "when we want to export message from a non-existing topic" do
-      before { get "explorer/messages/non-existing/0/1/export" }
-
-      it do
-        refute(response.ok?)
-        assert_equal(404, response.status)
-      end
-    end
-
-    context "when message exists" do
-      # Export deserializes the payload (default JSON deserializer) and re-serializes it with
-      # `#to_json`. A bare numeric payload like `rand.to_s` is not a JSON serialization fixed
-      # point: Ruby's `Float#to_s` and the json gem's float encoder occasionally disagree (e.g.
-      # scientific notation or precision), so `assert_equal(payload, response.body)` was flaky
-      # for roughly 1-in-2000 random values. A JSON string literal round-trips identically.
-      let(:payload) { rand.to_s.to_json }
-      let(:expected_file_name) { "#{topic}_0_0_payload.json" }
-      let(:expected_disposition) { "attachment; filename=\"#{expected_file_name}\"" }
-
-      before do
-        produce(topic, payload)
-        get "explorer/messages/#{topic}/0/0/export"
-      end
-
-      it do
-        assert_ok
-        assert_equal(expected_disposition, response.headers["content-disposition"])
-        assert_equal("application/octet-stream", response.headers["content-type"])
-        assert_equal(payload, response.body)
-      end
-    end
-
-    context "when message exists on a dynamic topic with custom deserializer" do
-      let(:payload) { rand.to_s }
-      let(:expected_file_name) { "#{topic}_0_0_payload.json" }
-      let(:expected_disposition) { "attachment; filename=\"#{expected_file_name}\"" }
-
-      before do
-        topic_name = topic
-
-        draw_routes do
-          pattern(/#{topic_name}/) do
-            active(false)
-            deserializer(->(_message) { "1" })
-          end
-        end
-
-        produce(topic, payload)
-        get "explorer/messages/#{topic}/0/0/export"
-      end
-
-      it "expect to use custom deserializer" do
-        assert_ok
-        assert_equal(expected_disposition, response.headers["content-disposition"])
-        assert_equal("application/octet-stream", response.headers["content-type"])
-        assert_equal('"1"', response.body)
-      end
-    end
-
-    context "when message exists but exports are off" do
-      let(:payload) { rand.to_s }
-
-      before do
-        Karafka::Web.config.ui.policies.messages.stubs(:export?).returns(false)
-
-        produce(topic, payload)
-        get "explorer/messages/#{topic}/0/0/export"
-      end
-
-      it do
-        refute(response.ok?)
-        assert_equal(403, response.status)
-      end
-    end
-
-    context "when message payload is JSON with non-UTF-8 encoded string values" do
-      before do
-        # JSON.parse accepts this payload but the resulting hash contains strings with invalid
-        # UTF-8 bytes, so serializing it back with #to_json raises JSON::GeneratorError. There
-        # is no JSON representation that could be exported, hence 404
-        produce(topic, Fixtures.binfile("payloads/invalid_utf8.bin"))
-        get "explorer/messages/#{topic}/0/0/export"
-      end
-
-      it do
-        refute(response.ok?)
-        assert_equal(404, response.status)
       end
     end
   end
