+++ test/timings/pro.json
@@ -18,7 +18,9 @@
   "test/lib/karafka/web/pro/ui/controllers/consumers/commanding_controller_test.rb": 12.894,
   "test/lib/karafka/web/pro/ui/controllers/consumers/jobs_controller_test.rb": 12.093,
   "test/lib/karafka/web/pro/ui/controllers/topics/configs_controller_test.rb": 11.491,
-  "test/lib/karafka/web/pro/ui/controllers/explorer/messages_controller_test.rb": 11.367,
+  "test/lib/karafka/web/pro/ui/controllers/explorer/messages/publishing_controller_test.rb": 8.0,
+  "test/lib/karafka/web/pro/ui/controllers/explorer/messages/republishing_controller_test.rb": 4.0,
+  "test/lib/karafka/web/pro/ui/controllers/explorer/messages/exporting_controller_test.rb": 4.0,
   "test/lib/karafka/web/pro/ui/controllers/scheduled_messages/schedules_controller_test.rb": 11.11,
   "test/lib/karafka/web/pro/ui/controllers/status_controller_test.rb": 9.888,
   "test/lib/karafka/web/pro/ui/controllers/topics/topics_controller_test.rb": 7.919,
